Add parametric option to bound_degree()
The parametric and non-parametric cases are all exactly identical for
the regular Risch Differential Equation and parametric Risch
Differential Equation, except for the former dc = c.degree(t) and in the
latter, dc = max([i.degree(t) for i in G]).

Other algorithms in prde.py are very similar to their counterparts in
rde.py, and could probably also benefit from such consolidation, though
none are as close as this one.