Use the backsubs dict in risch_integrate()
If we had to modify the integrand to integrate it, then we will try to
change it back after integrating.  For example,

Before:

In [1]: risch_integrate(diff(x**x, x), x)
Out[1]:
 x⋅log(x)
ℯ

After:

In [1]: risch_integrate(diff(x**x, x), x)
Out[1]:
 x
x

Note that right now this is very limited because of the
exp-exact-substitution hack.  In the future, this will also be used with
sin and cos, which must be converted to tan before integrating.