autowrap: allow ignoring OutputArguments when using 'args' keyword
While codegen is pedantic about missing arguments, the wrappers will convert
pure output arguments to return values.  It is inconsistent and confusing, if
the user is required to specify Output arguments that don't actually exist in
the function returned by autowrap.

Added tests of wrapped function arguments using the DummyWrapper.