Fix a bug in spde() in sympy/integrals/rde.py
quo is not the proper way to do a/b where a and b are Polys and
b.degree() == 0, because it only works when the ground domains of a and
b are fields.  Ideally, Poly would have a .mul_ground() method to make
this easier, but for now, just do Poly(a.as_basic/b.as_basic()) instead.