Add cancel_primitive() and primitive case in bound_degree() in Risch Algorithm
Also includes some other bug fixes.  This needs more tests, but so far,
this is a great improvement, and risch_integrate() is now able to do
many more integrals that previously raised NotImplementedError.  For
example:

In [1]: risch_integrate(diff(exp(2*x**2)*log(x)*log(x + 1), x), x)
Out[1]:
    2
 2⋅x
ℯ    ⋅log(x)⋅log(1 + x)

(by the way, it fails in the old integrate()).

More importantly, this is the last official commit for the 2010 Google
Summer of Code coding period.  I want to take this opportunity to thank
Google for having such a wonderful program, the Python Software
Foundation for accepting my project this year, and the SymPy development
team for their support, especially Chris Smith for his advice and my
mentor Mateusz Paprocki, because this project would have been impossible
without his excellent Polynomials module.

Note that even though this is the end of the official coding period for
GSoC, my project is far from over, and I plan on continuing to work on
it until it is completely implemented and merged, and I also plan on
doing more work on symbolic integration in SymPy beyond even that.