2027: don't combine negative bases under vanilla x
In a product of bases raised to the same power, a**x*b**x *...,
you can only combine bases that are positive if x is not an integer.

Also, a**(2*x) == (a**x)**2 but not vice versa unless a is positive
or x is an integer. Consider (-1)**(2*1/2) = (sqrt(-1))**2 = I**2 = -1
but sqrt((-1)**2) = sqrt(1) = 1. For that reason, the debate in test_pow2
was removed. None of those expressions should ever be true if the base is not
known to be positive.