secondquant: made AntiSymmetricTensor subclass of Expr
Before it was a subclass of Function, but that is inappropriate and
gives unnecessary overhead.  (Coupled-Cluster example with two-body
Hamiltonian runs ~5% faster without.)  It is also about 10% faster now
because we call _sort_anticommuting_fermions only twice, and avoid the
decorator.

Also promoted the eval docstring to a class docstring for easy access in
isympy