secondquant: Improve canonization of AntiSymmetricTensor
We choose to let dummy indices be sorted to the right and non-dummies to the
left.  A fixed order is necessary to generate correct keys for consistent dummy
substitution.  The keys must be 100% free from any dependence on the actual
dummy symbols in the expression.  Before this patch the generated key could be
different depending on whether a dummy happened to be to the left or right of
non-dummies in the AntiSymmetricTensor.

In addition, AntiSymmetricTensor will order indices with:

   particle < hole < general