1999 Integral.doit checks for equal limits
Although Integral is left unsimplified for a 0 integrand or identical
upper and lower limits (since Integral is suppose to be an unevaluated
object) a test for these conditions in doit were added and the
instantiation logic was refactored and the error message modified.

Tests were added to affirm that autosimplification is not done to
such integrals.