examples: Added Q.M. example with autowrapped numerical integration
The example demonstrates how the autowrap module can be used to generate binary
functions that perform a numerical approximation to an integral.  The
integration is used to find coefficients needed for the approximation of a
Hydrogen wave function as a linear combination of harmonic oscillator wave
functions.

Note: this example depends on numpy, and partly on pylab.