security: Fix a slightly dangerous file operation in test_autowrap.py
One of the tests used a call to autowrap with the tempdir keyword argument
specified as '/tmp/dummy'.  This would overwrite files in that directory
with the name wrapper_module_N.py and wrapper_module_N.pyc where N is a given
integer depending on the order of test execution.

It is highly unlikely that these files would exist on a system, unless they
have been created by test_autowrap.py.  So the danger is minimal.

This patch fixes the issue by removing the keyword argument in that test.