Removed custom QMul, QAdd and QPow usage.
This is a first attempt to remove our usage of custom QMul, QAdd
and QPow classes. For now it removes all of our validation logic,
but in the long run will allow the quantum stuff to better
integrate with the sympy core. This will also force us to
improve the core Mul, Add and Pow classes to better support
validation and combining of things.