Fixed some (but not all) of the problems in issue 2085
By changing the recursive call in heuristics() to call limit() instead,
this allows the intermediate steps [limit(1/x, x, 0, dir="+") in this
case] to be calculated correctly, and ensures that heuristics() is only
called as a last resort on these steps, which I presume was initially
desired.

Also added relevant tests.