2080 WildFunction and Dummy integration
WildFunction needs to reference Dummy.dummycount not Symbol.dummycount.
However, it might as well have its own indexing scheme. It now starts
from 0 rather than the current dummycount. In addition, WildFunctions
are not dummy entities, so it is perhaps misleading to be using the
dummycount index to provide a uniqe integer for the name.

Remove autonumbering of WildFunction since nothing else behaves
like that and it wasn't being used.

Add autonumbering to Dummy and remove Temporary (since Dummy now
will behave like that). If a temporary symbol is needed for something,
Dummy() can be used and a unique symbol will be created with name
being the current dummycount index prefaced (as usual) with an underscore:

>>> Dummy()
_0

dummycount was renamed count as it is always prefaced with Dummy in
order to change it; Dummy.dummycount seems redundant.