2095 fix Ana Flores solution of quartic
In the quasisymmetric solution of the quartic the coefficient d = (c/a)**2
is identified as m**2. The value of m was being computed as sqrt(d) rather
than c/a and this gives incorrect solutions. The test identifying that and
covering the rest of the quartic routine were added and minor edits made
throughout polyroots.py.

In addition, although I thought that there was no general solution to the
quartic (since it has several decisions that need to be made while
constructing the general solution) this does not seem to be the case.
Such comments were removed accordingly.

Finally, attemps to order the solutions was removed as in the end sets of
solutions will be of interest and one will likely have to test solutions
anyway.