2092 solve_linear added to solvers.py
This routine is based on the former solve4linearsymbol that was in
order.py since it will be useful there for eliminating linear expressions
from a poly system to decrease the number of non-linear equations
that needed to be solved.

Differentiation is used to find linear terms, but cancel() is not used so as
to ensure that the solution returned is truly linear and not, for example,
the solution of x=0 that would be obtained from x**2(1/x-y/x) which would
not satisfy (mathematically) the original expression which would involve
dividing by zero. In addition, solve_linear will solve for a linear symbol
that appears in the rhs argument -- this was something that the former
routine did not check for.

solve_linear also has the ability to attempt solutions for a symbol from
a given list while at the same time excluding solutions for certain variables.