2084 limit issues from the "-" direction
limit(1 + 1/x, x, 0, '-') was giving oo because it takes as the
unbounded result the value obtained when substituting x with 0
into the 1/x term. Instead, having quickly determined that the
term is unbounded (and not nan) it should then recalculate the
value from the desired direction.

In addition, limits involving powers evaluated from the negative
direction were added, i.e. 1/x**a with dir="-"

tests were added