Flatten internal format for integral limits.
Multidimensional limits used to be a tuple in args[1].  Now they are
stored as individual elements in args (that is, the list of limits is now args[1:])

For each dimension, the new format is also flat, and uses Tuple instead
of the built-in tuple.
The previous format used a nested tuple for the limits: (symbol, (lower,upper)).

Indefinite integration is represented by a Tuple of length 1: (symbol,)
Integration evaluated at a point is represented by a Tuple of length 2:
  (symbol, upper)
Definite integration is represented by a Tuple of length 3:
  (symbol, lower, upper)

Also add MathML output for sums, trivial now that the limits are stored in the
same format between Sum and Integral.