tsolve: modify a test to pass on 64-bit
This test was failing before on Python 2.7.1 64 bit.

A strange bug is triggered, there is an issue in distributing the minus sign.
The solution is to write

    -((numer)/den)

rather than

    -(numer)/den.

By putting (numer)/den in parenthesis, it is no longer an Add that the
-1 is multiplying, so you don't get the automatic distribution.

Thanks to Chris Smith for pointing this out!