Rename NonElementaryIntegral to NonElementaryIntegralException
This is because I will eventually create a subclass of Integral for
nonelementary integrals, and I plan on calling that
NonElementaryIntegral.

Note that I am calling this "Exception," not "Error" as exception
classes are often called, because when this exception is raised, it does
not signify an error, but rather it is simply a mechebism to jump out of
whatever subalgorithm it is in whenever it determines somehow that the
underlying integral is not elementary (think of it as a Pythonic goto
statement that isn't evil).