2124 nseries and lseries handle x0!=0
Series now should be more robust for more complex functions.

series() will call either nseries or lseries (when n=None) making
it possible to do either nseries or lseries from series.

   Note: It would also be possible to just make nseries and lseries
   helpers for series so there is only one "entry" to series; if the
   n=None option is not desired then iter=True could perhaps trigger
   the call to lseries. This could also respect the n value by returning
   n terms instead of calculating up to order n.)

The handling of direction and the value of x0 is handled in
series and series. All nseries, lseries, _eval_nseries and
_eval_lseries have to do is return the series term(s) as x -> 0
from the positive direction. For this reason, x0 values and are no
longer sent to the _eval routines.

getn was moved from power.py to expr.py and geto was changed to .getO()

Derivative.lseries now works; it had an unconditional stop in it before.

Issue 2123

If an expression has no order term then it should return right away
from lseries or else it will hang.

Issue 2122

Code in integrals.py intended for use by exponential.py was moved
to exponential.py and the general code for handling series terms
of an Integral left in integrals.py.