Create DifferentialExtension object and merge build_extension into it
This is just a C struct-like object that will hold all the information
about a differential extension so that the Risch Algorithm functions
don't have to pass it all around as several arguments.  This will be the
fourth and should hopefully be the last API change in the Risch
Algorithm

So far, I haven't changed the API of the Risch Algorithm functions to
use this directly, because I still need to create an easy way to create
this without having to build the extension from scratch for testing.
However, I have deleted the old build_extension(), and risch_integrate()
now calls this instead.

Also upgraded the old build_extension() tests to use this, and add some
new tests for it.