Add function as_poly_1t to handle K[t, 1/t] polynomials in the Risch Algorithm
This function is very hackish in that it uses internal Poly functions to
work, but it should be able to completely decide the problem of
converting a polynomial in K[t, 1/t] into a polynomial in K[t, z], where
z = 1/t, as opposed to the old usage of apart(), which might possibly
not work for some case (this should also be much faster than apart()).

See issue 2032 for more discussion on this.