raise doctest parse error
A bare exception was hiding the docstring parsing error which is raised
when a dectest is not whitespace terminated:

    Example:
        >>> print 1
	2
    This line should have a blank line before it or
    else it is interpreted as part of the doctest and is then
    flagged as having inconsistent whitespace.

Although the parser's regex can be fixed to detect such embedded tests,
tests really should have space after them for stylistic reasons.

There are 2 more bare exceptions in runtest.py.

A test that showed up as having problems after this fix was corrected.