2127 do not allow 0 or 1-arg AssocOp
Either case returns what it would have returned before, but now
it does so regardless of whether the evaluate is False. i.e. these
cases have higher precedence than the unevaluated expression.

The sympification of args was refactored, too, to do it once
before determining whether the expression is to be evaluated or not.