Make Tfuncs not reversed in the Risch Algorithm API
This was a stupid idea that I had, because Tfuncs needs to be
back-substituted in reverse order. But it turns out to just be a
headache and annoying to have it (and only it) in reverse order. It's
much easier to just call reversed() on Tfuncs before calling subs.  We
still do not include anything for x (T[0]) because zip() very nicely
truncates to the shortest list.

Note that I am also in the middle of changing the API of all the Risch
Algorithm functions to use DifferentialExtension. As a result, I was
unable to run the tests, so I am not certain that there are no silly
typos in this change.  If there are, then they will be fixed once I have
gotten everything into a testable state again.