Modifications to the cases attribute of DifferentialExtension
DifferentialExtension.cases is no longer reversed.  This is for the same
reason that DifferentialExtension.Tfuncs is no longer reversed:  in the
main use, it is reversed, but it makes things confusing to make it
reversed in the object itself.  Rather, it is much clearer to have
everything in the object be in the same order, and to just use
reversed() whenever you need something in the reverse order.

This also adds the attribute DifferentialExtension.case, which points to
DifferentialExtension.cases[DifferentialExtension.level], similar to
DifferentialExtension.t and DifferentialExtension.d.

Tests were added/corrected, but I am in the middle of changing all the
Risch Algorithm code to use DifferentialExtension. Therefore, it is
possible that tests do not pass with this change set, though unlikely.
If there are any problems, they will be fixed in a future commit.