Update mpmath in sympy to version 0.16
The only change from the original mpmath sources aside from the usual
conversions was to change an itertools.product(rows, columns) in
mpmath/matrices/matrices.py to ((a, b) for a in rows for b in columns).
This is because Python 2.5 and Python 2.4 do not have itertools.product.
According to the docstring of itertools.product, these are equivalent.

This does not yet update the mpmath docs.