Fix the doctests to run in Python 2.6.6 and Python 2.7.1
See issue 2041.  We had to copy over some more code from Python's
doctest.py into our own runtests.py to revert this Python commit
http://svn.python.org/view/python/branches/release26-maint/Lib/doctest.
py?r1=83492&r2=83493, which resets the sys.displayhook before running
the doctests.

We also had to copy over the function for testing .txt files so that it
could call the new modified SymPyDoctestRunner class, and we had to do
some mangling with a change in return type for that function between
Python 2.5 and Python 2.6.