Fix Issue 1947: @property methods are not doctested
The problem is that we are sorting our doctests by line number (the
Python doctest.py sorts by name) so when the _find_lineno function
couldn't find the line number for the docstring, we were just skipping
the docstring.

But @property docstrings are not found by this function.  For now, we
just replace the lineno with a dummy value in this case, but eventually
we will need to write our own version that is smarter, because it prints
an incorrect line number whenever a doctest fails in a @property
method's docstring.

This also fixes a handful of docstrings that are now caught.
Fortunately, there wasn't really anything that was incorrect, just
things that weren't imported in the doctest.