Allowed to use init_printing() in documentation
By default init_printing() stores configuration in Printer, using
set_global_settings() method. This is the best approach to make
all kinds of printers work configured in interactive sessions.
However, due to doctest's limitations, this approach fails in
the case of documentation. For this purpose, no_global option
was added to init_printing(). When set to True, it won't allow
init_printing() to store any global settings. In this scenario
only the chosen printer will be configured properly.