Rename sum() to summation()
This is because it conflicts with the builtin sum().  See issue 1376 and
issue 1727.

I also took the opportunity to fix Sum() so that it always remains
unevaluated by default and so that it and summation always require at
least symbols argument.  I have also written a docstring for
summation().

I fixed a test in test_pickling.py that I am not entirely sure if the
fix was correct (it was using the now defunct Sum(2) syntax).

There is a class called Sum2 that seems to duplicate summation/Sum.  I
did not look into it here.