tensor: allow Indexed with indices of any type
The Indexed constructor used to convert all indices to the Idx class,
that is confusing and also limits the possibilities.  This patch removes
this behaviour so that any expression can be used as index.

The docstrings are updated.  It is also made clear that only indices of
class Idx are subject to implicit summation.