always return Integral, even if trivial when differentiating
There are some obvious integral that can be done very quickly to
keep output clean when differentiating an Integral.
After this commit, no simplification will be done. The result, of course
can be simplified with doit(). But since this can be expensive, there should
perhaps be a simplification function for integrals which checks if integrals
which

- are zero because of limits being the same
- are zero because of integrand being zero
- have an integrand independent of integration variable
- have and integrand that is a polynomial in integration variable

Perhaps all this could be done by just giving doit() a keyword "easy"
to indicate that only the above integrals should be performed.

Here are some examples of output that are now different:

<This one has a 0 integrand>

    >>> i = Integral(x, (x, 5*y), (y, x, 2*x))
    >>> i.diff(x)
    -Integral(x, (x, 5*x)) + 2*Integral(x, (x, 10*x)) + \
    Integral(0, (y, x, 2*x))
was
    -Integral(x, (x, 5*x)) + 2*Integral(x, (x, 10*x))

<This one has an integrand free of the integration variable>

    >>> Integral(x, (a, 3*y)).diff(x)
    Integral(1, (a, 3*y))
was
    3*y

<This one has an integrand free of integration variable and two equal
limits, thus identically 0>

    >>> Integral(x + y, y, (y, 1, x)).diff(x)
    Integral(x + y, (y, x)) + Integral(Integral(1, (y, y)), (y, 1, x))
was
    Integral(y, (y, 1, x)) + Integral(x + y, (y, x))

<this one has an x-dependent integrand but identical limits and a 0
integrand>

    >>> i=Integral(x + y, (x, x, y), (y, y, x))
    >>> i.diff(x)
    Integral(2*x, (x, x, x)) + Integral(-x - y + \
    Integral(0, (x, x, y)), (y, y, x))
was
    Integral(-x - y, (y, y, x))

<this one has an x-independent integrand only>
    >>> Integral(y, y, x).diff(y)
    Integral(y, x)
was
    x*y