1983: unify mul and pow handling of Add**e
During the flatten procedure in Mul, code duplicates Add's
extraction of a root from a power that appears in _eval_power. It is
important that this code not be duplicated since any changes
to the one should really happen to the other. So now Mul
will call to Add to get the result.

The complicating issue is that _eval_power returns an expression
but Mul needs the two terms. So a keyword to allow the result to
come back as, e.g. (-1, 1+x) rather than -1-x has been added.

In addition, if coeff_terms is ever changed to give anything other
than +/-1 then the existing code will erroneously separate a negative
from a non-integer power so that has been fixed.

Conflicts:

	sympy/core/mul.py