2130: don't depend on subs in limit
2130 subs at discontinuity is NaN

    power subs gives NaN if either base or exponent are NaN
    or the substitution results in oo.

move find_limit's heuristic to limit

    All the heuristics except the one for log terms was already
    in limit so the log heuristic was moved to limit and find_limit
    was eliminated.

What constitutes a PoleError was expanded a bit. Since subs will give
a NaN now it is (more?) likely that a NaN will be returned than a
signed infinity. Also even though arg0 might not be zero the function
evaluated at arg0 may be undefined (as tan(x+pi/2) at x=0) and so the
PoleError is raised there, too.