tensor: Do not silently wrap indices as Idx
Before, the Indexed class would silently wrap its indices using the Idx
class.  This was confusing and unnecessary complex (issue 2059).   This
patch removes the silent wrapping, so that the user must explicitly
create Idx objects for the indices if that is desired.

This change also required different implementations of the .shape and
.ranges properties of Indexed.  A docstring was added for .shape.