tensor: Let Indexed.ranges return Tuple instead of tuple
The upper and lower ranges of an Indexed object used to be represented
in a tuple.  We now use a Tuple instead.

Rationale:
Since the upper and lower ranges may be any integer expression, it is
convenient to have the pair directly available as a Basic object.  For
example if a substitution is desired.