Make subs work with any iterable
It's assumed that list(sequence) will be of the form ([old1, new1], ...,
[oldn, newn]) when iterated (except for when sequence is a dict, which
is handled separately).

This uses hasattr(sequence, '__iter__'), which I hope is the best way to
check this.  Also, do not do type checking in _functions, and add a
test, cleaning up the test file a little bit too while we're at it.

Conflicts:

	sympy/core/tests/test_subs.py