Fixed logical error in Basic.free_symbols
In order to determine free and bound symbols of an expression tree, the
call to free_symbols must be forwarded to all Basic.args.  An integral
or a sum may be present at a deeper level.  Added a regression test and
a test for Basic.free_symbols.

Because ExprCondPair sometimes store a bool in self.args, it was
necessary to overload free_symbols from Basic.