Fixed logical error in Basic.free_symbols
In order to determine free and bound symbols of an expression tree, the
call to free_symbols must be forwarded to all Basic.args.  An integral
or a sum may be present at a deeper level.  Added a regression test and
a test for Basic.free_symbols.

Also, because ExprCondPair sometimes store a bool in self.args, it was
necessary to overload free_symbols from Basic. Both the expr and cond
(if it is a subclass of Basic) may have free_symbols.  Added a regression
test for this. (Thanks to Chris Smith <smichr@gmail.com> for pointing out
this problem.)