Fixed SMatrix transpose
Added functions transpose,row_list,col_list, and their respective property variables.
row_list is an alternate representation of Sparse Matrices, which is essentially a list of non-zero elements sorted by Row. Similarly for col_list.
The order for transpose is O(L) where L is the number of non-zero elements, as opposed to the naive representations's O(n**2)