2061 as_coeff defines coefficient as leading Rational
Issue 2061 identified a problem inherent in how Reals are treated
differently than NumberSymbols. Whereas (pi*x).as_coeff_mul() gives 1,
(pi, x) (3.0*x).as_coeff_mul() gave 3.0, (x,). This is a problem because
(based on no failing tests after the change) sympy is not ever expecting
to pull out anything but a Rational with as_coeff_*.

Number now checks to see if the coeff is a Rational and if not, only
extracts the sign of the Number as occurs with expressions like pi*x and
pi + x which would return 1 and 0 from as_coeff_mul and as _coeff_add,
respectively.