Changes in the printing logic of physics.quantum.
in InnerProduct:
- The printing of the bra in InnerProduct must contain all the information
and not only the quantum numbers (i.e. the label). Now _print_content is used
instead of _print_label so all the info is printed instead of only the
quantum numbers.

in QExpr(and side effects in other classes):
- The printing logic was changed
-- The code is potentially shorter as now less subclasses will be
required to mess around with brackets.
-- The brackets stuff is done once and for all in the State class.
-- The _print_contents methods in all children will be required to
append some info to _print_label and not to add brackets.
- Side effects
-- The changes are mostly invisible to the subclasses, but now it's easier
to implement the new printing methods for Spin.
-- The printing of InnerProduct was refactored and is more straightforward
and symmetrical in respect to the bra and the ket.

in Spin:
- Making the printing of J?Kets and Bras less ambiguous.