Fix issue 1715.
sympy/series/gruntz.py
  (mrv_leadterm) If mrv returns set([]), do not rewrite.
sympy/series/tests/test_gruntz.py
  (test_issue1715) Add test for the issue that triggered this.

Note:
The problem in mrv_leadterm is that the function mrv does more
simplifications, so even though mrv_leadterm may not have concluded
that the expression is independent of x, mrv may, and then return an
empty set. mrv_leadterm needs to handle this.
The example in question is (x + 1)**(1/log(x + 1)), which is actually
constantly E for positive x.