Updated examples in docstrings for symarray in matrices.py
The examples will pass under doctest but they are skipped since it symarray
requires numpy to work. Issue 2252 has some discussion on how to get the
examples to work with or without numpy but they are too convoluted and make
the examples hard to read. Since they are examples I consider it is more
important for them to be readable than to pass under doctest on a pure python
env.