2179: Store Lambda variables in a Tuple.
The variables of the Lambda must be specified as part of a tuple. The syntax
Lambda(x, y, z, x*y*z) is disallowed, Lambda((x, y, z), x*y*z) should be
used instead. This simplifies the signature and the internal structure
of Lambda.

As a side-effect, nullary Lambdas are now allowed and
sympify('lambda : 1') == Lambda((), 1).