Polynomial division will default to a field
Over a field (QQ), the following example gives non-trivial quotient:

In [1]: div(x**2 + 1, 2*x - 4)
Out[1]:
⎛    x   ⎞
⎜1 + ─, 5⎟
⎝    2   ⎠

However, over a ring the quotient is is zero:

In [2]: div(x**2 + 1, 2*x - 4, domain=ZZ)
Out[2]:
⎛        2⎞
⎝0, 1 + x ⎠

In [3]: div(x**2 + 1, 2*x - 4, auto=False)
Out[3]:
⎛        2⎞
⎝0, 1 + x ⎠

Previously div() didn't automatically switch to a field, which
was impractical and caused trouble when div() was used as a
sub-algorithm in other modules.

To make div() not use automatic conversion, either specify
a domain explicitly or set 'auto' flag to False. The same
applies to rem(), quo() and exquo().