Rename as_coeff_Mul to as_coeff_mul(args=False) (issue 2307)
The former was not used very often, and was similar to the latter except
the latter returns the args of the non-coeff part and the former returns
a Mul.  Furthermore, as_coeff_mul(), whereas as_coeff_Mul() was billed
only as being an efficient way to extract a coefficient.  Also,
as_coeff_Mul() was only being used in four places in the code, one of
which should really have been using as_coeff_mul() (it was just
splitting apart the args again).

The tests for as_coeff_Mul() were converted to tests for
as_coeff_mul(args=False).

There is a test failure due to a significant difference between the two:

In [1]: (4.0*x).as_coeff_mul()
Out[1]: (1, (4.0, x))

In [2]: (4.0*x).as_coeff_Mul()
Out[2]: (4.0, x)

The latter behavior is only needed in one place (_pi_coeff() in
sympysympy/functions/elementary/trigonometric.py), so it needs to be
emulated.  I haven't fixed this yet.