Implement randomised testing.
As requested in issues 2281 and 1746, this adds a parameter --seed
to bin/test which
can be used to specify the PRNG seed. The implementation in runtest.py will
re-seed the PRNG with this value at the beginning of testing each file.
The seed also determines the order of tests when running with --random.
If no seed is passed a random one is generated.
The seed is output at the beginning of the tests, right after ground types.

Also adds helper functions for randomised testing, in the new file randtest.py.

Usages of this will be added in the next commits.