Implement randomised testing.
As requested in issue 2281, this adds a parameter --seed to bin/test which
can be used to specify the PRNG seed. The implementation in runtest.py will
re-seed the PRNG with this value at the beginning of testing each file.
If no seed is passed a random one is used.
If there are test failures, the seed is output.

Also adds helper functions for randomised testing, in the new file randtest.py.

Usages of this will be added in the next commits.

bin/test: Add option "seed".
sympy/utilities/runtests.py:
  (test) Accept a seed, create a random one if needed. Pass seed on to tester.
  (SymPyTests.__init__) Accept and store a seed.
  (SymPyTests.test) Pass seed to reporter.
  (SymPyTests.test_file) Re-seed PRNG.
  (PyTestReporter.finish) Print seed if there are test failures.

sympy/utilities/randtest.py: New file.