Revamp bessel type functions.
This commit reorganises the bessel functions.

There is a new abstract class BesselBase. The specific functions besselj,
bessely, besselk, besseli, hankel1 and hankel2 derive from it.
Differentiation, some rewriting, and latex printing are implemented.
For the latter, there is a new helper function _print_BesselBase in
printing/latex.py.

There is also a new base class SphericalBesselBase, from which yn and jn
derive. This replaces the old functions, which directly returned the
expressions in terms of trigonometric and rational functions. In order
to get such an expression for integral orders, use expand_func().

All new classes also have docstrings outlining the definitons of the objects
involved, and giving references to literature.

There is a general TODO note at the top of the file.