2307: remove as_coeff_Mul
    If these are ever really deemed necessary the following could
    be used:

    >>> def as_coeff0_mul(eq):
    ...  if eq.is_Mul:
    ...    c, x = eq.as_two_terms()
    ...    if c.is_Number:
    ...      return c, x
    ...  elif eq.is_Number:
    ...    return eq, S.One
    ...  return S.One, eq
    ...
    >>> def as_coeff0_add(eq):
    ...  if eq.is_Add:
    ...    c, x = eq.as_two_terms()
    ...    if c.is_Number:
    ...      return c, x
    ...  elif eq.is_Number:
    ...    return eq, S.Zero
    ...  return S.Zero, eq

If more than just Numbers are to be collected then all arguments have to be searched since a number (as opposed to Number) might appear anywhere, e.g.

    >>> 3*x*(1+sqrt(2))
    3*x*(1 + 2**(1/2))

The number that is the Add appears as the last argument above.
To find the number portion of an expression, e.as_independent(Symbol)
could be used which, for the above, gives (3 + 3*2**(1/2), x).