Improve changing between xyz bases for spin states
Use the Wigner D-function to define the represent function for spin
states which allows for statesof any j and m values to be represented as
vectors in the new basis.

Implementation of rewrite, which is used to write states in terms of
different basis is improved to work for all j and m values and for
changing to any basis, not just to z. This utilizes the represent
function.

Changed the InnerProduct evaluation logic to use the represent functions
allowing for innerproducts to be evaluated between any spin states.