Raise NotImplementedError for f(g(x)).diff(x)
This require derivatives evaluated at a point to really be correct.  The
previous Derivative(f(g(x)), g(x))*Derivative(g(x), x) was a hack that
lead to problems (for example, if you tried to run .doit() on that, it
would fail).  It was difficult to continue to support this and to not
allow erroneous input to diff (like f(x).diff(2)) because of the way the
code is implemented with the chain rul.  Since the output was
technically incorrect anyway, this was removed.

Note that this is a workaround (though in my opinion, the old way was
too, and it was more hackish).  The correct fix would be to implemented
derivatives evaluated at a point (issue 1620).

I had to wrap some code in classify_ode() around a try, except
NotImplementedError, block to make it work with things like f(y/x) in an
ode doctest (the classification test was for 1st_exact, but the doctest
related to homogeneous_coeff and wasn't exact).  Other code might have
to do similar things until this gets fixed properly (though no other
code in the library code needed to be fixed).

See issue 1525 for more information on this issue.