Create class for Clebsch-Gordan and Wigner-3j coeffs
This creates classes implementeing symbolic Clebsch-Gordan coefficients
and symbolic Wigner-3j symbols. These coefficients can be created and
manipulated as expressions. When numerical values are given for the
terms, they can be evaluated using a doit() method.

This also begins to implement the simplification of Clebsch-Gordan
coefficients by the symmetries in the coefficients with the cg_simp
function.