Improve find_instantiations
find_instantiations is a function which, given two sets of hypergeometric
parameters (class IndexPair) ip and nip, tries to guess values for the free
symbols of nip, such that ip is reachable from the substituted nip.
This is used to implement parametric hypergeometric formulae.

This can now handle all cases in the tests, but it is still far from perfect.
It is also (predictably) quite slow. But for now it is almost never the bottleneck.

sympy/simplify/hyperexpand.py:
  (IndexPair.reachable) Rename to ...
  (IndexPair.difficulty) ... this. Also compute difficulty.
  (Formula.find_instantiations) Try all permutations.
              Also try to be clever about numerators.
              Also try some origins nearby, they might be quicker.
              Finally try to make integers positive.
  (Formula.is_suitable) Throw out whacky instantiations with nan or oo.
  (FormulaCollection.lookup_origin) Look for nearest origin.